Refactor: Change Night from Module to superclass
