+++ lib/night_writer.rb
@@ -2,18 +2,21 @@ require 'pry'
 require './lib/english_translator'
 require './lib/night_helper'
 
-class NightWriter
-  include NightHelper
-  attr_reader :file, :read_message, :new_file
+class NightWriter < Night
+  # include NightHelper
+  # attr_reader :file, :read_message, :new_file
 
-  def initialize(file)
-    @incoming_file = file
-    @read_message = read_file.strip
-    @new_file = create_new_file
+  # def initialize(file)
+  #   @incoming_file = file
+  #   @read_message = read_file.strip
+  #   @new_file = create_new_file
+  #   @translator = EnglishTranslator.new(@read_message)
+  #   run
+  # end
+
+  def create_translator
     @translator = EnglishTranslator.new(@read_message)
-    run
   end
-
   # def run
   #   write_to_new_file
   #   print_update
