--- new_message.txt
@@ -1,1 +1,0 @@
-one bright day in the middle of the night two dead boys got up to fight back to back they faced each other drew their swords and apples
